refactor: common 패키지 분리
