+++ src/main/java/kitchenpos/common/infra/DefaultPurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.infra;
+package kitchenpos.common.infra;
 
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.stereotype.Component;
@@ -6,6 +6,8 @@ import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriComponentsBuilder;
 
 import java.net.URI;
+
+import kitchenpos.common.domain.PurgomalumClient;
 
 @Component
 public class DefaultPurgomalumClient implements PurgomalumClient {
